python: make ply a soft dependency for JobspecV1
Problem: Importing JobspecV1 unconditionally imports ply via the
flux.constraint.parser module, even when the constraint parser is
never used.

Move MiniConstraintParser from flux.job._utils into
flux.constraint.parser, removing the top-level ConstraintParser import
from _utils.py.  Replace the top-level ConstraintSyntaxError and
MiniConstraintParser imports in Jobspec.py with a lazy import inside
the apply_options() --requires handling block, so ply is only loaded
when constraint parsing is actually needed.
